Connect Home to store and get mockEvents
