+++ Actions/index.js
@@ -1,0 +1,4 @@
+export const getEvents = events => ({
+  type: 'GET_EVENTS',
+  events
+})
