+++ Reducers/EventReducer.js
@@ -1,0 +1,8 @@
+export const EventReducer = (state = [], action) => {
+  switch (action.type) {
+    case 'GET_EVENTS':
+      return action.events
+    default:
+      return state
+  }
+}
