+++ Reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import { EventReducer } from './EventReducer'
+
+export const rootReducer = combineReducers({
+  events: EventReducer
+})
