+++ package.json
@@ -22,7 +22,9 @@
     "react-redux": "^7.1.1",
     "redux": "^4.0.4",
     "redux-axios-middleware": "^4.0.1",
+    "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
+    "remote-redux-devtools": "^0.5.16",
     "styled-components": "^4.3.2",
     "yarn": "^1.17.3"
   },
